requirements.txt
groq==0.3.0
pyqt5==5.15.9
speechrecognition==3.10.0
pygame==2.5.2
gtts==2.3.2
pyttsx3==2.90
python-dotenv==1.0.0
watchdog==3.0.0